refactor: Split window.rs into modular structure
- Created 7 logical modules:
  * partitioning.rs - Partition management
  * sorting.rs - Partition sorting and value comparison
  * frames.rs - Frame boundary calculation
  * ranking.rs - ROW_NUMBER, RANK, DENSE_RANK, NTILE
  * aggregates.rs - COUNT, SUM, AVG, MIN, MAX window functions
  * value.rs - LAG, LEAD functions
  * utils.rs - Shared utilities
  * tests.rs - All unit tests (46 tests)
- Converted window.rs to mod.rs with re-exports
- All 46 window function tests passing
- No functional changes, pure refactoring

Reduces window.rs from 1,609 lines to ~40 lines mod.rs + 8 focused modules
