+++ crates/executor/src/evaluator/window/partitioning.rs
@@ -1,0 +1,69 @@
+//! Row partitioning for window functions
+//!
+//! Groups rows into partitions based on PARTITION BY expressions.
+
+use ast::Expression;
+use storage::Row;
+use types::SqlValue;
+
+/// A partition of rows for window function evaluation
+#[derive(Debug, Clone)]
+pub struct Partition {
+    pub rows: Vec<Row>,
+}
+
+impl Partition {
+    pub fn new(rows: Vec<Row>) -> Self {
+        Self { rows }
+    }
+
+    pub fn len(&self) -> usize {
+        self.rows.len()
+    }
+
+    pub fn is_empty(&self) -> bool {
+        self.rows.is_empty()
+    }
+}
+
+/// Partition rows by PARTITION BY expressions
+///
+/// Groups rows into partitions based on partition expressions.
+/// If no PARTITION BY clause, all rows go into a single partition.
+pub fn partition_rows<F>(
+    rows: Vec<Row>,
+    partition_by: &Option<Vec<Expression>>,
+    eval_fn: F,
+) -> Vec<Partition>
+where
+    F: Fn(&Expression, &Row) -> Result<SqlValue, String>,
+{
+    // If no PARTITION BY, return all rows in single partition
+    let Some(partition_exprs) = partition_by else {
+        return vec![Partition::new(rows)];
+    };
+
+    if partition_exprs.is_empty() {
+        return vec![Partition::new(rows)];
+    }
+
+    // Group rows by partition key values (use BTreeMap for deterministic ordering)
+    let mut partitions_map: std::collections::BTreeMap<Vec<String>, Vec<Row>> =
+        std::collections::BTreeMap::new();
+
+    for row in rows {
+        // Evaluate partition expressions for this row
+        let mut partition_key = Vec::new();
+
+        for expr in partition_exprs {
+            let value = eval_fn(expr, &row).unwrap_or(SqlValue::Null);
+            // Convert to string for grouping (handles NULL consistently)
+            partition_key.push(format!("{:?}", value));
+        }
+
+        partitions_map.entry(partition_key).or_default().push(row);
+    }
+
+    // Convert HashMap to Vec<Partition>
+    partitions_map.into_values().map(Partition::new).collect()
+}
